api/requirements.txt
black
coverage
fastapi
flake8
loguru
pytest
python-dotenv
sqlalchemy
python-telegram-bot
tinvest